+++ service/src/main/kotlin/io/provenance/explorer/domain/models/explorer/CommonModels.kt
@@ -70,29 +70,6 @@ data class CountStrTotal(
     val denom: String
 )
 
-data class TokenSupply(
-    val currentSupply: CoinStr,
-    val circulation: CoinStr,
-    val communityPool: CoinStr,
-    val bonded: CoinStr
-)
-
-data class TokenDistributionPaginatedResults(
-    val ownerAddress: String,
-    val data: CountStrTotal
-)
-
-data class TokenDistributionAmount(
-    val denom: String,
-    val amount: String
-)
-
-data class TokenDistribution(
-    val range: String,
-    val amount: TokenDistributionAmount,
-    val percent: String
-)
-
 enum class Timeframe { WEEK, DAY, HOUR, FOREVER }
 
 const val hourlyBlockCount = 720